#MX-748 In Code Review
Update he module go.mod
restructure th eproject to remove the internal directory
