+++ go.mod
@@ -1,18 +1,10 @@
 module github.com/messagex/go-messagex
 
-go 1.13
+go 1.14
 
 require (
-	github.com/davecgh/go-spew v1.1.0
-	github.com/gogs/chardet v0.0.0-20191104214054-4b6791f73a28 // indirect
-	github.com/jaytaylor/html2text v0.0.0-20200412013138-3577fbdbcff7 // indirect
-	github.com/jhillyerd/enmime v0.8.0 // indirect
-	github.com/joho/godotenv v1.3.0 // indirect
-	github.com/mattn/go-runewidth v0.0.9 // indirect
-	github.com/olekukonko/tablewriter v0.0.4 // indirect
-	github.com/pkg/errors v0.9.1 // indirect
 	github.com/rs/zerolog v1.18.0
 	github.com/stretchr/testify v1.5.1
-	golang.org/x/net v0.0.0-20200506145744-7e3656a0809f // indirect
-	golang.org/x/tools v0.0.0-20200513022744-65e69ff2d148 // indirect
+	golang.org/x/net v0.0.0-20200513185701-a91f0712d120 // indirect
+	golang.org/x/tools v0.0.0-20200513201620-d5fe73897c97 // indirect
 )
