+++ messagex.go
@@ -6,10 +6,10 @@
 package messagex
 
 import (
-	"github.com/messagex/go-messagex/internal/pkg/messagexapi"
-	"github.com/messagex/go-messagex/internal/types/api"
-	"github.com/messagex/go-messagex/internal/types/constants"
 	"github.com/messagex/go-messagex/pkg/logger"
+	"github.com/messagex/go-messagex/pkg/messagexapi"
+	"github.com/messagex/go-messagex/types/api"
+	"github.com/messagex/go-messagex/types/constants"
 )
 
 type APIClient struct {
