+++ messagex_test.go
@@ -1,14 +1,14 @@
 package messagex
 
 import (
+	"github.com/messagex/go-messagex/pkg/logger"
+	"github.com/messagex/go-messagex/pkg/messagexapi"
+	"github.com/messagex/go-messagex/types/api"
+	"github.com/messagex/go-messagex/types/constants"
+	"github.com/messagex/go-messagex/util/mocks"
+	"github.com/messagex/go-messagex/util/testdata"
 	"github.com/stretchr/testify/assert"
 	"testing"
-	"github.com/messagex/go-messagex/internal/pkg/messagexapi"
-	"github.com/messagex/go-messagex/internal/types/api"
-	"github.com/messagex/go-messagex/internal/types/constants"
-	"github.com/messagex/go-messagex/internal/util/mocks"
-	"github.com/messagex/go-messagex/internal/util/testdata"
-	"github.com/messagex/go-messagex/pkg/logger"
 )
 
 var l *logger.Logger
