+++ types/constants/constants.go
@@ -12,7 +12,4 @@ const (
 	// Endpoints
 	EndpointAuthorise = "authorise"
 	EndpointMailSend = "mail/send"
-
-	// Other
-	AttachmentType = "attachment"
 )
